Add configurable size `Table` structure
A table is defined by a number of "buckets", B, a number of "slots" S,
and a maximum range of buckets a source may occupy W. When we creating
bitcoin connections, we hear about new peers from a "source", who gives
us a number of "destinations". The table is a 2D matrix, where each
gossiped network address derives an (x, y) location based the source and
destination. Given the source, we first limit the number of potential
buckets they may add network addresses to. This is deterministic based
on their "/16". We pick a final bucket using some entropy from the
destination. Then the destination "/16" is used to compute the slot
within the bucket. If there is already a network address at that
(bucket, slot), we return the conflicting network address for testing.

On conflicts the user can either: call `is_terrible` to determine if it
should be dropped, or establish a short-lived connection to test for
liveliness.

When selecting a peer, we do so by picking a random bucket. If that
bucket is empty, we iterate over the buckets, overwise return a random
slot in the bucket.

The same source adding the same network address should always conflict,
and if we have at least one peer, selecting a peer should always work.